Issue #33, #41, #49, #74 partially
