+++ cbam/cbam/doctype/customs_import/customs_import.js
@@ -11,7 +11,7 @@ frappe.ui.form.on("Customs Import", {
 	refresh(frm) {
 		frm.add_custom_button(__("Sent all Emails (yet not sent)"), function () {
 			frappe.call({
-				method: "cbam.detail_confirmation_email.send_email", // OPEN
+				method: "cbam.send_email_from_customs_import.create_user_and_send_email", // OPEN
 				args: {
 					goods_list: cur_frm.doc.goods,
 				},
