+++ cbam/cbam/doctype/supplier_employee/supplier_employee.js
@@ -6,3 +6,18 @@
 
 // 	},
 // });
+
+
+frappe.ui.form.on("Supplier Employee", {
+	refresh(frm) {
+		frm.add_custom_button(__("Sent Email"), function () {
+			frappe.call({
+				method: "cbam.send_email_from_employee.create_user_and_send_email", // OPEN
+				args: {
+					employee: cur_frm.docname,
+                    supplier: cur_frm.doc.supplier_company,
+				},
+			});
+		}, __("⚠️ Look out! ⚠️"));
+	},
+});
