+++ cbam/fixtures/role_profile.json
@@ -7,6 +7,9 @@
   "role_profile": "00 Reporting Declarant",
   "roles": [
    {
+    "parent": "00 Reporting Declarant",
+    "parentfield": "roles",
+    "parenttype": "Role Profile",
     "role": "Reporting Declarant"
    }
   ]
@@ -19,6 +22,9 @@
   "role_profile": "00 Supplier",
   "roles": [
    {
+    "parent": "00 Supplier",
+    "parentfield": "roles",
+    "parenttype": "Role Profile",
     "role": "Supplier"
    }
   ]
